Added analysis to simulation
